+++ pkg/scaling/scalers_builder.go
@@ -258,6 +258,8 @@ func buildScaler(ctx context.Context, client client.Client, triggerType string, 
 		return scalers.NewSplunkScaler(config)
 	case "stan":
 		return scalers.NewStanScaler(config)
+	case "sumologic":
+		return scalers.NewSumologicScaler(config)
 	case "temporal":
 		return scalers.NewTemporalScaler(ctx, config)
 	default:
